backend/requirements.txt
certifi==2017.7.27.1
Django==1.11.3
django-bootstrap3==9.1.0
djangorestframework==3.7.1
pytz==2017.2

superuser
user: anthonychen
pw: B******
